refactor: Update imports and services for domain restructuring
- Update imports in all controllers, repositories, services, mappers
- Maintain existing functionality after moving user entities
- Fix all compilation errors from package restructuring
