+++ drumigo/src/main/java/com/ftn/drumigo/domain/DriverProfileChangeRequest.java
@@ -1,6 +1,8 @@
 package com.ftn.drumigo.domain;
 
 import com.ftn.drumigo.domain.enums.RequestStatus;
+import com.ftn.drumigo.domain.users.Admin;
+import com.ftn.drumigo.domain.users.Driver;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
