+++ drumigo/src/main/java/com/ftn/drumigo/domain/Message.java
@@ -1,6 +1,7 @@
 package com.ftn.drumigo.domain;
 
 import com.ftn.drumigo.domain.enums.MessageType;
+import com.ftn.drumigo.domain.users.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
