+++ drumigo/src/main/java/com/ftn/drumigo/domain/Notification.java
@@ -1,6 +1,7 @@
 package com.ftn.drumigo.domain;
 
 import com.ftn.drumigo.domain.enums.NotificationType;
+import com.ftn.drumigo.domain.users.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
