+++ drumigo/src/main/java/com/ftn/drumigo/domain/Review.java
@@ -1,5 +1,7 @@
 package com.ftn.drumigo.domain;
 
+import com.ftn.drumigo.domain.users.Driver;
+import com.ftn.drumigo.domain.users.Passenger;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
