+++ drumigo/src/main/java/com/ftn/drumigo/domain/UserNote.java
@@ -1,5 +1,7 @@
 package com.ftn.drumigo.domain;
 
+import com.ftn.drumigo.domain.users.Admin;
+import com.ftn.drumigo.domain.users.User;
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
