--- drumigo/src/main/java/com/ftn/drumigo/dto/DriverDocumentResponse.java
@@ -1,11 +1,0 @@
-package com.ftn.drumigo.dto;
-
-import java.time.Instant;
-
-public record DriverDocumentResponse(
-        Long id,
-        Long driverId,
-        String documentName,
-        String documentUrl,
-        Instant uploadedAt) {
-}
