+++ drumigo/src/main/java/com/ftn/drumigo/dto/VehicleResponse.java
@@ -9,13 +9,10 @@ public record VehicleResponse(
     String driverSurname,
     Long vehicleTypeId,
     String vehicleTypeName,
-    String model,
-    String licensePlate,
     Integer numSeats,
     Boolean babyFriendly,
     Boolean petFriendly,
     BigDecimal currentLat,
-    BigDecimal currentLng,
-    Boolean available
+    BigDecimal currentLng
 ) {}
