+++ drumigo/src/main/java/com/ftn/drumigo/event/RideFinishedEventListener.java
@@ -1,7 +1,7 @@
 package com.ftn.drumigo.event;
 
 import com.ftn.drumigo.domain.Notification;
-import com.ftn.drumigo.domain.Passenger;
+import com.ftn.drumigo.domain.users.Passenger;
 import com.ftn.drumigo.domain.Ride;
 import com.ftn.drumigo.domain.RidePassenger;
 import com.ftn.drumigo.domain.RideWaypoint;
