+++ drumigo/src/main/java/com/ftn/drumigo/mapper/VehicleMapper.java
@@ -18,14 +18,11 @@ public class VehicleMapper {
             vehicle.getDriver() != null ? vehicle.getDriver().getSurname() : null,
             vehicle.getVehicleType() != null ? vehicle.getVehicleType().getId() : null,
             vehicle.getVehicleType() != null ? vehicle.getVehicleType().getName().name() : null,
-            vehicle.getModel(),
-            vehicle.getLicensePlate(),
             vehicle.getNumSeats(),
             vehicle.getBabyFriendly(),
             vehicle.getPetFriendly(),
             vehicle.getCurrentLat(),
-            vehicle.getCurrentLng(),
-            vehicle.getAvailable()
+            vehicle.getCurrentLng()
         );
     }
 }
