+++ drumigo/src/main/java/com/ftn/drumigo/repository/AdminRepository.java
@@ -1,6 +1,6 @@
 package com.ftn.drumigo.repository;
 
-import com.ftn.drumigo.domain.Admin;
+import com.ftn.drumigo.domain.users.Admin;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
