+++ drumigo/src/main/java/com/ftn/drumigo/repository/FavoriteRouteRepository.java
@@ -1,7 +1,7 @@
 package com.ftn.drumigo.repository;
 
 import com.ftn.drumigo.domain.FavoriteRoute;
-import com.ftn.drumigo.domain.Passenger;
+import com.ftn.drumigo.domain.users.Passenger;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
