+++ drumigo/src/main/java/com/ftn/drumigo/repository/PassengerRepository.java
@@ -1,6 +1,6 @@
 package com.ftn.drumigo.repository;
 
-import com.ftn.drumigo.domain.Passenger;
+import com.ftn.drumigo.domain.users.Passenger;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
