+++ drumigo/src/main/java/com/ftn/drumigo/repository/VehicleRepository.java
@@ -8,8 +8,6 @@ import java.util.List;
 
 @Repository
 public interface VehicleRepository extends JpaRepository<Vehicle, Long> {
-    List<Vehicle> findByAvailableTrue();
-    boolean existsByLicensePlate(String licensePlate);
-    java.util.Optional<Vehicle> findByDriver(com.ftn.drumigo.domain.Driver driver);
+    java.util.Optional<Vehicle> findByDriver(com.ftn.drumigo.domain.users.Driver driver);
 }
