+++ drumigo/src/main/java/com/ftn/drumigo/service/DriverService.java
@@ -1,20 +1,17 @@
 package com.ftn.drumigo.service;
 
-import com.ftn.drumigo.domain.Driver;
-import com.ftn.drumigo.domain.DriverDocument;
+import com.ftn.drumigo.domain.users.Driver;
 import com.ftn.drumigo.domain.UserToken;
 import com.ftn.drumigo.domain.Vehicle;
 import com.ftn.drumigo.domain.VehicleType;
 import com.ftn.drumigo.domain.enums.TokenType;
 import com.ftn.drumigo.domain.enums.UserRole;
 import com.ftn.drumigo.dto.DriverCreateRequest;
-import com.ftn.drumigo.dto.DriverDocumentCreateRequest;
 import com.ftn.drumigo.dto.DriverUpdateRequest;
 import com.ftn.drumigo.dto.SetPasswordRequest;
 import com.ftn.drumigo.exception.BadRequestException;
 import com.ftn.drumigo.exception.ConflictException;
 import com.ftn.drumigo.exception.ResourceNotFoundException;
-import com.ftn.drumigo.repository.DriverDocumentRepository;
 import com.ftn.drumigo.repository.DriverRepository;
 import com.ftn.drumigo.repository.UserRepository;
 import com.ftn.drumigo.repository.UserTokenRepository;
@@ -39,18 +36,12 @@ public class DriverService {
     private final UserRepository userRepository;
     private final VehicleRepository vehicleRepository;
     private final VehicleTypeRepository vehicleTypeRepository;
-    private final DriverDocumentRepository driverDocumentRepository;
     private final UserTokenRepository userTokenRepository;
     
     public Driver create(DriverCreateRequest request) {
         // Check if email already exists
         if (userRepository.existsByEmail(request.email())) {
             throw new ConflictException("User with email " + request.email() + " already exists");
-        }
-        
-        // Check if license plate already exists
-        if (vehicleRepository.existsByLicensePlate(request.vehicleLicensePlate())) {
-            throw new ConflictException("Vehicle with license plate " + request.vehicleLicensePlate() + " already exists");
         }
         
         // Get vehicle type
@@ -64,9 +55,6 @@ public class DriverService {
         driver.setEmail(request.email());
         driver.setAddress(request.address());
         driver.setPhone(request.phone());
-        driver.setLicenseNumber(request.licenseNumber());
-        driver.setRole(UserRole.DRIVER);
-        driver.setActive(false); // Not active until password is set
         driver.setActiveDriver(false);
         driver.setBlocked(false);
         driver.setCreatedAt(Instant.now());
@@ -78,12 +66,9 @@ public class DriverService {
         Vehicle vehicle = new Vehicle();
         vehicle.setDriver(driver);
         vehicle.setVehicleType(vehicleType);
-        vehicle.setModel(request.vehicleModel());
-        vehicle.setLicensePlate(request.vehicleLicensePlate());
         vehicle.setNumSeats(request.vehicleNumSeats());
         vehicle.setBabyFriendly(request.vehicleBabyFriendly() != null ? request.vehicleBabyFriendly() : false);
         vehicle.setPetFriendly(request.vehiclePetFriendly() != null ? request.vehiclePetFriendly() : false);
-        vehicle.setAvailable(true);
         
         vehicleRepository.save(vehicle);
         
@@ -122,33 +107,13 @@ public class DriverService {
         if (request.phone() != null) {
             driver.setPhone(request.phone());
         }
-        if (request.licenseNumber() != null) {
-            driver.setLicenseNumber(request.licenseNumber());
-        }
-        
         driver.setUpdatedAt(Instant.now());
         
         return driverRepository.save(driver);
     }
     
-    public DriverDocument createDocument(Long driverId, DriverDocumentCreateRequest request) {
-        Driver driver = getById(driverId);
-        
-        DriverDocument document = new DriverDocument();
-        document.setDriver(driver);
-        document.setDocumentName(request.documentName());
-        document.setDocumentUrl(request.documentUrl());
-        document.setUploadedAt(Instant.now());
-        
-        return driverDocumentRepository.save(document);
-    }
-    
     public String createActivationToken(Long driverId) {
         Driver driver = getById(driverId);
-        
-        if (driver.getActive()) {
-            throw new BadRequestException("Driver is already active");
-        }
         
         // Generate token
         String token = UUID.randomUUID().toString();
@@ -183,7 +148,6 @@ public class DriverService {
 
         Driver driver = (Driver) userToken.getUser();
         driver.setPasswordHash(passwordHash);
-        driver.setActive(true);
         driver.setUpdatedAt(Instant.now());
         
         driverRepository.save(driver);
