+++ drumigo/src/main/java/com/ftn/drumigo/service/ReportService.java
@@ -1,6 +1,6 @@
 package com.ftn.drumigo.service;
 
-import com.ftn.drumigo.domain.Passenger;
+import com.ftn.drumigo.domain.users.Passenger;
 import com.ftn.drumigo.domain.Ride;
 import com.ftn.drumigo.domain.RidePassenger;
 import com.ftn.drumigo.domain.enums.RideStatus;
